"1:50pm.
```
The learning rate is 2 ** -15
At iteration 9 the cost and the sign accuracy are 24.43139 and 0.998/1.
At iteration 19 the cost and the sign accuracy are 19.57848 and 0.994/1.
At iteration 29 the cost and the sign accuracy are 10.17217 and 0.988/1.
At iteration 39 the cost and the sign accuracy are 3.932954 and 0.988/1.
At iteration 49 the cost and the sign accuracy are 4.033669 and 0.986/1.
At iteration 59 the cost and the sign accuracy are 4.819073 and 0.986/1.
At iteration 69 the cost and the sign accuracy are 6.276945 and 0.956/1.
At iteration 79 the cost and the sign accuracy are 5.022727 and 0.97/1.
At iteration 89 the cost and the sign accuracy are 7.037267 and 0.962/1.
At iteration 99 the cost and the sign accuracy are 3.074681 and 0.998/1.
At iteration 109 the cost and the sign accuracy are 6.210053 and 0.964/1.
At iteration 119 the cost and the sign accuracy are 5.586793 and 0.964/1.
At iteration 129 the cost and the sign accuracy are 4.985634 and 0.972/1.
At iteration 139 the cost and the sign accuracy are 2.367488 and 0.992/1.
At iteration 149 the cost and the sign accuracy are 4.961555 and 0.978/1.
At iteration 159 the cost and the sign accuracy are 3.743879 and 0.988/1.
At iteration 169 the cost and the sign accuracy are 5.34921 and 0.976/1.
At iteration 179 the cost and the sign accuracy are 2.925123 and 0.99/1.
At iteration 189 the cost and the sign accuracy are 3.739558 and 0.986/1.
At iteration 199 the cost and the sign accuracy are 3.184916 and 0.99/1.
At iteration 209 the cost and the sign accuracy are 2.118681 and 0.988/1.
At iteration 219 the cost and the sign accuracy are 5.836059 and 0.964/1.
At iteration 229 the cost and the sign accuracy are 6.323936 and 0.968/1.
At iteration 239 the cost and the sign accuracy are 64.23544 and 0.66/1.
At iteration 249 the cost and the sign accuracy are 82.79979 and 0.586/1.
At iteration 259 the cost and the sign accuracy are 79.19981 and 0.604/1.
At iteration 269 the cost and the sign accuracy are 95.59978 and 0.522/1.
At iteration 279 the cost and the sign accuracy are 87.99977 and 0.56/1.
At iteration 289 the cost and the sign accuracy are 73.59982 and 0.632/1.
At iteration 299 the cost and the sign accuracy are 77.19982 and 0.614/1.
At iteration 309 the cost and the sign accuracy are 83.19981 and 0.584/1.
At iteration 319 the cost and the sign accuracy are 70.79984 and 0.646/1.
At iteration 329 the cost and the sign accuracy are 73.19984 and 0.634/1.
At iteration 339 the cost and the sign accuracy are 78.79982 and 0.606/1.
At iteration 349 the cost and the sign accuracy are 52.39988 and 0.738/1.
At iteration 359 the cost and the sign accuracy are 84.3998 and 0.578/1.
At iteration 369 the cost and the sign accuracy are 70.39983 and 0.648/1.
At iteration 379 the cost and the sign accuracy are 85.19978 and 0.574/1.
At iteration 389 the cost and the sign accuracy are 82.7998 and 0.586/1.
At iteration 399 the cost and the sign accuracy are 57.59986 and 0.712/1.
At iteration 409 the cost and the sign accuracy are 61.19987 and 0.694/1.
At iteration 419 the cost and the sign accuracy are 73.59982 and 0.632/1.
At iteration 429 the cost and the sign accuracy are 88.39981 and 0.558/1.
At iteration 439 the cost and the sign accuracy are 97.59976 and 0.512/1.
At iteration 449 the cost and the sign accuracy are 85.19981 and 0.574/1.
At iteration 459 the cost and the sign accuracy are 71.19982 and 0.644/1.
At iteration 469 the cost and the sign accuracy are 61.59986 and 0.692/1.
At iteration 479 the cost and the sign accuracy are 60.39984 and 0.698/1.
At iteration 489 the cost and the sign accuracy are 70.39983 and 0.648/1.
At iteration 499 the cost and the sign accuracy are 78.7998 and 0.606/1.
```

The result I get with map is completely different.

Could it be that `ln_tanh` is broken somehow?

1:55pm.

```
                    //ln_tanh {alpha plastic static=input}
                    map CudaAD.Activation.hebb_tanh {alpha plastic static=input}
```

The discrepancy between these two is fairly insane and that is not even getting into the difference between the old and this RNN which should be identical without KFAC.

I need to investigate this. So many unexpected bugs cropping up. I need to squish them all."