"2:35pm.
```
[|
    [|[-1.98, -1.98]|]
|]

[|
    [|[0, 0]|]
|]

[|
    [|[0, 0]|]
|]

[|
    [|[0, 0]|]
|]

[|
    [|[0, 0]|]
|]

[|
    [|[0, 0]|]
|]

[|
    [|[0.01, -1.98]; [0.01, -1.98]|]
|]
```

Hmm, only a single gradient on the final step.

```
[|
    [|[-9.9, 0]|]
|]

[|
    [|[0, 0]|]
|]

[|
    [|[0, 0]|]
|]

[|
    [|[0, 0]|]
|]

[|
    [|[0, 0]|]
|]

[|
    [|[0, 0]|]
|]

[|
    [|[0.01, -9.9]; [0.01, 0]|]
|]
```

Ohhh, you have to be kidding me. Both of them only have gradients at the final step. Should there not be some kind of propagation going on? Let me print out the `out`. That one should be the same for both."