"3:25pm. Ah, I just had to turn off the update. Now I can investigate this properly.
```
{bodies={ar=var (array_cuda_global (float32)); block=[]; offset=lit 0i64; size=[lit 1i64]}; dim=[lit 1i64]}
```

This is the size of a.

```
inl a = a i
```

This is how I am indexing into it.

3:30pm.

```
The learning rate is 2 ** -10
[|
    [|[-9.9, 0]|]
|]

[|
    [|[0, 0]|]
|]

[|
    [|[0.01, -9.9]; [0.01, 0]|]
|]
```

It was a out of bounds error. I just did not notice it because it never occurred to me to test whether it was happening.

FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF...

I need to repeat all the experiments done with weight norm.

3:35pm. Obviously, I am not going to do that now, but let me sure that the results between the two functions are equal now.

Agh, it was one little weakness, just one little slip in the armor. Well, thankfully I discovered it. Only the `wn_hebb` should be affected by this. Almost definitely the reason why I could not get a net with two parameters to optimize with SGD is because of this.

Those earlier plastic RNN architecture with modulation might start working once I get back to them now.

3:40pm. Ok...

This is actually rather good news everything considered. I thought it was my ideas that were wrong, but it turns out there was a bug in the program. KFAC certainly is not going to beat that.

...No wait. Am I rotating the inputs or the outputs in wn_hebb?

When I printed out the atomic_add there were a whole bunch of threads...

Ahhh, I see. It needs to sum up the adjoints for the output. Yeah, the conclusion that I need to repeat the tests still stands.

3:45pm. Ready, go. It is time that I start bringing things into line."