"3:55pm. Yeah, they are equal. Let me see if I can get it to converge even with modulation.
```
The learning rate is 2 ** -18
At iteration 9 the cost and the sign accuracy are 25.29767 and 0.998/1.
At iteration 19 the cost and the sign accuracy are 24.06834 and 1/1.
At iteration 29 the cost and the sign accuracy are 21.28435 and 0.992/1.
At iteration 39 the cost and the sign accuracy are 14.56956 and 1/1.
At iteration 49 the cost and the sign accuracy are 6.528305 and 0.988/1.
At iteration 59 the cost and the sign accuracy are 5.020621 and 0.986/1.
At iteration 69 the cost and the sign accuracy are 3.849248 and 0.972/1.
At iteration 79 the cost and the sign accuracy are 5.493985 and 0.982/1.
At iteration 89 the cost and the sign accuracy are 7.791356 and 0.976/1.
At iteration 99 the cost and the sign accuracy are 3.402177 and 0.998/1.
At iteration 109 the cost and the sign accuracy are 5.623434 and 0.972/1.
At iteration 119 the cost and the sign accuracy are 5.035214 and 0.97/1.
At iteration 129 the cost and the sign accuracy are 4.914778 and 0.976/1.
At iteration 139 the cost and the sign accuracy are 2.427714 and 0.994/1.
At iteration 149 the cost and the sign accuracy are 3.174294 and 0.988/1.
At iteration 159 the cost and the sign accuracy are 16.53702 and 0.972/1.
At iteration 169 the cost and the sign accuracy are 14.04051 and 0.984/1.
At iteration 179 the cost and the sign accuracy are 4.950175 and 0.998/1.
At iteration 189 the cost and the sign accuracy are 3.997411 and 0.986/1.
At iteration 199 the cost and the sign accuracy are 3.374488 and 0.984/1.
At iteration 209 the cost and the sign accuracy are 2.305412 and 0.988/1.
At iteration 219 the cost and the sign accuracy are 2.917899 and 0.984/1.
At iteration 229 the cost and the sign accuracy are 2.225286 and 0.994/1.
At iteration 239 the cost and the sign accuracy are 6.249475 and 0.976/1.
At iteration 249 the cost and the sign accuracy are 4.22124 and 0.978/1.
At iteration 259 the cost and the sign accuracy are 2.974761 and 0.996/1.
At iteration 269 the cost and the sign accuracy are 4.569719 and 0.976/1.
At iteration 279 the cost and the sign accuracy are 3.216062 and 0.982/1.
At iteration 289 the cost and the sign accuracy are 4.084062 and 0.984/1.
At iteration 299 the cost and the sign accuracy are 2.036355 and 0.996/1.
At iteration 309 the cost and the sign accuracy are 4.7886 and 0.978/1.
At iteration 319 the cost and the sign accuracy are 2.679553 and 0.998/1.
At iteration 329 the cost and the sign accuracy are 3.691484 and 0.988/1.
At iteration 339 the cost and the sign accuracy are 3.657214 and 0.978/1.
At iteration 349 the cost and the sign accuracy are 4.97123 and 0.976/1.
At iteration 359 the cost and the sign accuracy are 4.144414 and 0.988/1.
At iteration 369 the cost and the sign accuracy are 1.970438 and 0.996/1.
At iteration 379 the cost and the sign accuracy are 2.722051 and 0.982/1.
At iteration 389 the cost and the sign accuracy are 3.062905 and 0.986/1.
At iteration 399 the cost and the sign accuracy are 6.212301 and 0.966/1.
At iteration 409 the cost and the sign accuracy are 4.12423 and 0.982/1.
At iteration 419 the cost and the sign accuracy are 4.630426 and 0.984/1.
At iteration 429 the cost and the sign accuracy are 9.208333 and 0.936/1.
At iteration 439 the cost and the sign accuracy are 3.431551 and 0.994/1.
At iteration 449 the cost and the sign accuracy are 2.299242 and 0.994/1.
At iteration 459 the cost and the sign accuracy are 4.416362 and 0.982/1.
At iteration 469 the cost and the sign accuracy are 5.943223 and 0.976/1.
At iteration 479 the cost and the sign accuracy are 2.962891 and 0.988/1.
At iteration 489 the cost and the sign accuracy are 2.64014 and 0.988/1.
At iteration 499 the cost and the sign accuracy are 3.653373 and 0.982/1.
```

Yeah, it works. Phew.

Let me try it with KFAC.

4pm.

```
Starting timing for: Training
The learning rate is 2 ** -13
At iteration 9 the cost and the sign accuracy are 10.71657 and 0.962/1.
At iteration 19 the cost and the sign accuracy are 4.149707 and 0.986/1.
At iteration 29 the cost and the sign accuracy are 5.497321 and 0.966/1.
At iteration 39 the cost and the sign accuracy are 3.448142 and 0.988/1.
At iteration 49 the cost and the sign accuracy are 4.079189 and 0.986/1.
At iteration 59 the cost and the sign accuracy are 4.615634 and 0.986/1.
At iteration 69 the cost and the sign accuracy are 3.305243 and 0.974/1.
At iteration 79 the cost and the sign accuracy are 4.37048 and 0.978/1.
At iteration 89 the cost and the sign accuracy are 6.700881 and 0.978/1.
At iteration 99 the cost and the sign accuracy are 3.948534 and 0.998/1.
At iteration 109 the cost and the sign accuracy are 5.459007 and 0.978/1.
At iteration 119 the cost and the sign accuracy are 4.887285 and 0.97/1.
At iteration 129 the cost and the sign accuracy are 4.649682 and 0.988/1.
At iteration 139 the cost and the sign accuracy are 2.858885 and 1/1.
At iteration 149 the cost and the sign accuracy are 3.031163 and 0.99/1.
At iteration 159 the cost and the sign accuracy are 3.086625 and 0.988/1.
At iteration 169 the cost and the sign accuracy are 5.206355 and 0.972/1.
At iteration 179 the cost and the sign accuracy are 2.846495 and 0.99/1.
At iteration 189 the cost and the sign accuracy are 3.286512 and 0.99/1.
At iteration 199 the cost and the sign accuracy are 3.21673 and 0.99/1.
At iteration 209 the cost and the sign accuracy are 2.545553 and 0.988/1.
At iteration 219 the cost and the sign accuracy are 2.742659 and 0.992/1.
At iteration 229 the cost and the sign accuracy are 2.205454 and 0.996/1.
At iteration 239 the cost and the sign accuracy are 4.569376 and 0.98/1.
At iteration 249 the cost and the sign accuracy are 4.236142 and 0.972/1.
At iteration 259 the cost and the sign accuracy are 2.639106 and 0.996/1.
At iteration 269 the cost and the sign accuracy are 4.401504 and 0.976/1.
At iteration 279 the cost and the sign accuracy are 3.078212 and 0.98/1.
At iteration 289 the cost and the sign accuracy are 3.957309 and 0.984/1.
At iteration 299 the cost and the sign accuracy are 2.038441 and 0.996/1.
At iteration 309 the cost and the sign accuracy are 4.500883 and 0.978/1.
At iteration 319 the cost and the sign accuracy are 2.395148 and 0.998/1.
At iteration 329 the cost and the sign accuracy are 3.81654 and 0.99/1.
At iteration 339 the cost and the sign accuracy are 3.487137 and 0.978/1.
At iteration 349 the cost and the sign accuracy are 4.834817 and 0.974/1.
At iteration 359 the cost and the sign accuracy are 4.171462 and 0.988/1.
At iteration 369 the cost and the sign accuracy are 2.248941 and 0.996/1.
At iteration 379 the cost and the sign accuracy are 2.877958 and 0.984/1.
At iteration 389 the cost and the sign accuracy are 2.932093 and 0.986/1.
At iteration 399 the cost and the sign accuracy are 5.163675 and 0.97/1.
At iteration 409 the cost and the sign accuracy are 3.695886 and 0.982/1.
At iteration 419 the cost and the sign accuracy are 3.445433 and 0.984/1.
At iteration 429 the cost and the sign accuracy are 9.379549 and 0.93/1.
At iteration 439 the cost and the sign accuracy are 3.244508 and 0.994/1.
At iteration 449 the cost and the sign accuracy are 2.699612 and 0.996/1.
At iteration 459 the cost and the sign accuracy are 3.624341 and 0.982/1.
At iteration 469 the cost and the sign accuracy are 5.556963 and 0.978/1.
At iteration 479 the cost and the sign accuracy are 3.250938 and 0.99/1.
At iteration 489 the cost and the sign accuracy are 2.57907 and 0.994/1.
At iteration 499 the cost and the sign accuracy are 3.567778 and 0.984/1.
The time was 00:00:07.1713732 for: Training
```

Here is KFAC. Noted that with SGD, it only starts working with...-16. KFAC is carrying its weight for once."