"4pm. Now...when did I first make the Seq module. How many days ago was it?
From two days ago meaning 11/2 is when I need to go back through the commits and repeat all the experiments.

Well, for the time being, how about I rerun the experiment with the RNN like this one except with split biases?

I am actually quite tired now, but I think I can go on for a bit more. I'll leave reruning the experiments for tomorrow. The last two days were just so disappointing - I had great ideas that did not pan out, but now it looks like they might. I hope it will not turn out to be that again. I do not need a humble pie twice.

4:10pm. Let me start. Let me gather my strength and do one more experiment in order to make sure that I am sane.

4:20pm.

```
The learning rate is 2 ** -16
At iteration 9 the cost and the sign accuracy are 22.92161 and 0.998/1.
At iteration 19 the cost and the sign accuracy are 6.001213 and 0.986/1.
At iteration 29 the cost and the sign accuracy are 6.222319 and 0.97/1.
At iteration 39 the cost and the sign accuracy are 7.287025 and 0.982/1.
At iteration 49 the cost and the sign accuracy are 17.61243 and 0.988/1.
At iteration 59 the cost and the sign accuracy are 11.0874 and 0.986/1.
At iteration 69 the cost and the sign accuracy are 12.00894 and 0.986/1.
At iteration 79 the cost and the sign accuracy are 11.54858 and 0.988/1.
At iteration 89 the cost and the sign accuracy are 8.233454 and 0.962/1.
At iteration 99 the cost and the sign accuracy are 4.028186 and 0.99/1.
At iteration 109 the cost and the sign accuracy are 6.81029 and 0.964/1.
At iteration 119 the cost and the sign accuracy are 5.821448 and 0.968/1.
At iteration 129 the cost and the sign accuracy are 14.78269 and 0.974/1.
At iteration 139 the cost and the sign accuracy are 11.77096 and 1/1.
At iteration 149 the cost and the sign accuracy are 5.045556 and 0.988/1.
At iteration 159 the cost and the sign accuracy are 8.8592 and 0.964/1.
At iteration 169 the cost and the sign accuracy are 6.830789 and 0.972/1.
At iteration 179 the cost and the sign accuracy are 3.828962 and 0.986/1.
At iteration 189 the cost and the sign accuracy are 12.55707 and 0.982/1.
At iteration 199 the cost and the sign accuracy are 14.81712 and 0.996/1.
At iteration 209 the cost and the sign accuracy are 2.170569 and 0.988/1.
At iteration 219 the cost and the sign accuracy are 3.172525 and 0.984/1.
At iteration 229 the cost and the sign accuracy are 2.847583 and 0.994/1.
At iteration 239 the cost and the sign accuracy are 7.870924 and 0.966/1.
At iteration 249 the cost and the sign accuracy are 4.255254 and 0.984/1.
At iteration 259 the cost and the sign accuracy are 4.09093 and 0.99/1.
At iteration 269 the cost and the sign accuracy are 4.638049 and 0.972/1.
At iteration 279 the cost and the sign accuracy are 4.45143 and 0.986/1.
At iteration 289 the cost and the sign accuracy are 5.375024 and 0.984/1.
At iteration 299 the cost and the sign accuracy are 3.078386 and 0.996/1.
At iteration 309 the cost and the sign accuracy are 8.413012 and 0.978/1.
At iteration 319 the cost and the sign accuracy are 3.182812 and 0.994/1.
At iteration 329 the cost and the sign accuracy are 3.970505 and 0.988/1.
At iteration 339 the cost and the sign accuracy are 4.156701 and 0.984/1.
At iteration 349 the cost and the sign accuracy are 22.18515 and 0.99/1.
At iteration 359 the cost and the sign accuracy are 20.62491 and 0.992/1.
At iteration 369 the cost and the sign accuracy are 2.277365 and 0.996/1.
At iteration 379 the cost and the sign accuracy are 3.093077 and 0.982/1.
At iteration 389 the cost and the sign accuracy are 3.716424 and 0.986/1.
At iteration 399 the cost and the sign accuracy are 8.00502 and 0.972/1.
At iteration 409 the cost and the sign accuracy are 7.008922 and 0.982/1.
At iteration 419 the cost and the sign accuracy are 9.632765 and 0.984/1.
At iteration 429 the cost and the sign accuracy are 22.22683 and 0.968/1.
At iteration 439 the cost and the sign accuracy are 25.64753 and 1/1.
At iteration 449 the cost and the sign accuracy are 15.06852 and 1/1.
At iteration 459 the cost and the sign accuracy are 10.48928 and 0.978/1.
At iteration 469 the cost and the sign accuracy are 7.743329 and 0.98/1.
At iteration 479 the cost and the sign accuracy are 2.989339 and 0.99/1.
At iteration 489 the cost and the sign accuracy are 3.967316 and 0.988/1.
At iteration 499 the cost and the sign accuracy are 4.32017 and 0.974/1.
The time was 00:00:04.7588779 for: Training
```

Unfortunately, the two curves are not the same. Let me turn on the asserts.

...Though, do they have to be the same? I am amazed that atomics do not make the training non-deterministic."