"9:25am. https://www4.gogoanime.in/category/gyakuten-saiban-sono-shinjitsu-igi-ari
The music I've been listening to is making me want to pick this up. The concert by Tokyo Philharmonic Orchestra is godly.

9:30am. At any rate, let me start.

I need to start with a commit. I've figured out everything during the night. Basically the function I was looking for is `exp << tanh`.

Rather than have it taper off, what I will have it do is have it determine a linear distance in the log domain from the upper learning rate. By that I mean...

```
exp(x log upper + (1 - x) log lower)
```

`x` here could be the result of a matrix multiply. With tanh it could even go in the other direction. With both horizontal and vertical modulation it could be turned into...

```
inl x = (a + b) / 2
exp(x log upper + (1 - x) log lower)
```

The most important concept to understand here is `exp << tanh` function. It is symmetric around one in the log domain and bounded.

Originally I wanted to make the gating a parameter `m` and update it like `m' = m * exp (tanh x)`, but I think what I described above should suffice too and will have less hyperparameters for me to deal with. The activation will stabilize the network and make modulation viable.

I am thinking of making the upper limit just 1, or maybe I could make it something like 0.2. Yeah, 0.2 seems like it would be fine. And the lower limit should be something like 0.001. On the Binary Pattern, the important thing for me is that the damned training finally stabilizes in the presence of modulation. I can leave the extreme fine tuning to later. I can even make the `lower` and `upper` trainable parameters.

9:55am. Also...`exp(x log upper + (1 - x) log lower)` is just sigmoid cross entropy it seems. Well, now I know what it does now that I've rederived it from first principles on my own.

Ah, actually, no, the cross entropy is mixing the `exp` at the end. And in CE, the network is not calculating `x`, but `upper` and `lower`.

10am. Agh, damn the soundtrack. I want to watch the new Spyce ep, but because it is so good, I cannot bear to turn it off. I guess I'll start programming then.

First I will just check whether it works with standard modulation and then I will bring back the net that does it fully."