"11am.
```
Starting timing for: Training
The learning rate is 2 ** -18
At iteration 9 the cost and the sign accuracy are 26.21079 and 0.998/1.
At iteration 19 the cost and the sign accuracy are 25.7972 and 1/1.
At iteration 29 the cost and the sign accuracy are 25.0301 and 0.996/1.
At iteration 39 the cost and the sign accuracy are 23.43291 and 1/1.
At iteration 49 the cost and the sign accuracy are 19.70614 and 0.992/1.
At iteration 59 the cost and the sign accuracy are 11.62323 and 0.99/1.
At iteration 69 the cost and the sign accuracy are 4.298987 and 0.984/1.
At iteration 79 the cost and the sign accuracy are 4.849193 and 0.98/1.
At iteration 89 the cost and the sign accuracy are 7.267172 and 0.968/1.
At iteration 99 the cost and the sign accuracy are 3.201278 and 0.996/1.
At iteration 109 the cost and the sign accuracy are 5.63337 and 0.972/1.
At iteration 119 the cost and the sign accuracy are 5.069869 and 0.97/1.
At iteration 129 the cost and the sign accuracy are 4.931675 and 0.976/1.
At iteration 139 the cost and the sign accuracy are 2.426575 and 0.996/1.
At iteration 149 the cost and the sign accuracy are 3.19856 and 0.988/1.
At iteration 159 the cost and the sign accuracy are 17.2526 and 0.972/1.
At iteration 169 the cost and the sign accuracy are 15.95054 and 0.984/1.
At iteration 179 the cost and the sign accuracy are 5.853815 and 0.998/1.
At iteration 189 the cost and the sign accuracy are 3.969404 and 0.99/1.
At iteration 199 the cost and the sign accuracy are 3.406881 and 0.984/1.
At iteration 209 the cost and the sign accuracy are 2.321657 and 0.988/1.
At iteration 219 the cost and the sign accuracy are 2.875692 and 0.984/1.
At iteration 229 the cost and the sign accuracy are 2.310367 and 0.994/1.
At iteration 239 the cost and the sign accuracy are 6.324105 and 0.976/1.
At iteration 249 the cost and the sign accuracy are 4.13975 and 0.978/1.
At iteration 259 the cost and the sign accuracy are 2.894991 and 0.996/1.
At iteration 269 the cost and the sign accuracy are 4.598859 and 0.976/1.
At iteration 279 the cost and the sign accuracy are 3.165596 and 0.982/1.
At iteration 289 the cost and the sign accuracy are 4.182327 and 0.978/1.
At iteration 299 the cost and the sign accuracy are 2.021543 and 0.996/1.
At iteration 309 the cost and the sign accuracy are 4.697323 and 0.978/1.
At iteration 319 the cost and the sign accuracy are 2.721402 and 0.998/1.
At iteration 329 the cost and the sign accuracy are 3.783232 and 0.988/1.
At iteration 339 the cost and the sign accuracy are 3.655296 and 0.978/1.
At iteration 349 the cost and the sign accuracy are 6.010074 and 0.976/1.
At iteration 359 the cost and the sign accuracy are 4.218329 and 0.988/1.
At iteration 369 the cost and the sign accuracy are 1.980434 and 0.996/1.
At iteration 379 the cost and the sign accuracy are 2.813489 and 0.982/1.
At iteration 389 the cost and the sign accuracy are 3.046804 and 0.986/1.
At iteration 399 the cost and the sign accuracy are 6.28756 and 0.966/1.
At iteration 409 the cost and the sign accuracy are 4.176538 and 0.982/1.
At iteration 419 the cost and the sign accuracy are 4.615473 and 0.984/1.
At iteration 429 the cost and the sign accuracy are 9.353751 and 0.936/1.
At iteration 439 the cost and the sign accuracy are 3.481056 and 0.994/1.
At iteration 449 the cost and the sign accuracy are 2.366972 and 0.984/1.
At iteration 459 the cost and the sign accuracy are 4.370787 and 0.98/1.
At iteration 469 the cost and the sign accuracy are 5.953666 and 0.976/1.
At iteration 479 the cost and the sign accuracy are 2.998642 and 0.988/1.
At iteration 489 the cost and the sign accuracy are 2.649158 and 0.988/1.
At iteration 499 the cost and the sign accuracy are 3.686012 and 0.982/1.
The time was 00:00:07.6352011 for: Training
```

It does seem like it works. I forgot how it went last time.

11:05am. Oh, the above was with KFAC. It blows up with lower LRs, but does well until that point. Let me try making upper and lower hyperparameters."