"11:25am.
```
The learning rate is 2 ** -17
At iteration 9 the cost and the sign accuracy are 26.06683 and 0.998/1.
At iteration 19 the cost and the sign accuracy are 24.67794 and 1/1.
At iteration 29 the cost and the sign accuracy are 18.15227 and 0.992/1.
At iteration 39 the cost and the sign accuracy are 6.703436 and 0.988/1.
At iteration 49 the cost and the sign accuracy are 9.191272 and 0.986/1.
At iteration 59 the cost and the sign accuracy are 5.541891 and 0.986/1.
At iteration 69 the cost and the sign accuracy are 3.52267 and 0.98/1.
At iteration 79 the cost and the sign accuracy are 5.813807 and 0.982/1.
At iteration 89 the cost and the sign accuracy are 7.248069 and 0.962/1.
At iteration 99 the cost and the sign accuracy are 3.374976 and 0.99/1.
At iteration 109 the cost and the sign accuracy are 5.66381 and 0.968/1.
At iteration 119 the cost and the sign accuracy are 5.294999 and 0.968/1.
At iteration 129 the cost and the sign accuracy are 5.577109 and 0.976/1.
At iteration 139 the cost and the sign accuracy are 2.257346 and 0.996/1.
At iteration 149 the cost and the sign accuracy are 4.044856 and 0.988/1.
At iteration 159 the cost and the sign accuracy are 21.08452 and 0.97/1.
At iteration 169 the cost and the sign accuracy are 24.87891 and 0.99/1.
At iteration 179 the cost and the sign accuracy are 18.89141 and 0.998/1.
At iteration 189 the cost and the sign accuracy are 5.615985 and 0.99/1.
At iteration 199 the cost and the sign accuracy are 3.733635 and 0.984/1.
At iteration 209 the cost and the sign accuracy are 2.576958 and 0.988/1.
At iteration 219 the cost and the sign accuracy are 2.90807 and 0.984/1.
At iteration 229 the cost and the sign accuracy are 2.646814 and 0.994/1.
At iteration 239 the cost and the sign accuracy are 6.07612 and 0.974/1.
At iteration 249 the cost and the sign accuracy are 3.891016 and 0.974/1.
At iteration 259 the cost and the sign accuracy are 3.035769 and 0.996/1.
At iteration 269 the cost and the sign accuracy are 4.599967 and 0.976/1.
At iteration 279 the cost and the sign accuracy are 3.470399 and 0.986/1.
At iteration 289 the cost and the sign accuracy are 4.361883 and 0.978/1.
At iteration 299 the cost and the sign accuracy are 2.100723 and 0.996/1.
At iteration 309 the cost and the sign accuracy are 4.983664 and 0.976/1.
At iteration 319 the cost and the sign accuracy are 3.039629 and 0.998/1.
At iteration 329 the cost and the sign accuracy are 3.890079 and 0.988/1.
At iteration 339 the cost and the sign accuracy are 3.649209 and 0.982/1.
At iteration 349 the cost and the sign accuracy are 23.58396 and 0.978/1.
At iteration 359 the cost and the sign accuracy are 24.76539 and 0.996/1.
At iteration 369 the cost and the sign accuracy are 18.13029 and 1/1.
At iteration 379 the cost and the sign accuracy are 4.599767 and 0.984/1.
At iteration 389 the cost and the sign accuracy are 3.208483 and 0.986/1.
At iteration 399 the cost and the sign accuracy are 6.6259 and 0.966/1.
At iteration 409 the cost and the sign accuracy are 4.910115 and 0.982/1.
At iteration 419 the cost and the sign accuracy are 6.608844 and 0.984/1.
At iteration 429 the cost and the sign accuracy are 11.95197 and 0.946/1.
At iteration 439 the cost and the sign accuracy are 3.519002 and 0.994/1.
At iteration 449 the cost and the sign accuracy are 2.715453 and 0.986/1.
At iteration 459 the cost and the sign accuracy are 5.058495 and 0.98/1.
At iteration 469 the cost and the sign accuracy are 6.314216 and 0.976/1.
At iteration 479 the cost and the sign accuracy are 2.974468 and 0.988/1.
At iteration 489 the cost and the sign accuracy are 3.543642 and 0.99/1.
At iteration 499 the cost and the sign accuracy are 3.893161 and 0.974/1.
```

This is without KFAC. This is not too bad for this problem.

...KFAC makes no difference whatsover here. All the action is in the biases. Maybe if I applied KFAC to the baises I could increse the LR to -14 or -15, but it does not manner. For realistic tasks there won't be such a huge emphasis on learning the 3 biases.

11:30am.

```
                inm {eta out} =
                    inm data = matmult_stream {weights with data}
                    succ (wrap_split ((), inner.weights.static) data)
                inm out =
                    inm plastic = matmult_stream {data weight={T=H}; streams=weights.streams; block=()}
```

Let me merge the two multiplies together.

11:30am.

```
    inl plastic_rnn size =
        {
        init = inl sublayer_size ->
            open Initializer.dual.TensorView
            inl outer = {bias=1; input=sublayer_size; state=size}
            inl inner =
                {
                weights =
                    {
                    static={eta=1; out=size}
                    plastic=size
                    }
                biases =
                    {
                    alpha = 1
                    upper = 1
                    lower = 1
                    }
                }
            inl init =
                {
                weights =
                    {
                    bias={eta=const zero; out=const zero}
                    input={eta=const zero; out=identity}
                    state={eta=const zero; out=const zero}
                    }
                biases =
                    {
                    alpha = const 0.01f32
                    upper = const 0.1f32
                    lower = const 0.001f32
                    }
                }
            {
            dsc =
                {
                weights = weight {init=init.weights; dim=outer,inner.weights.static}
                biases = bias {init=init.biases; dim=1,inner.biases}
                streams = stream, stream
                dim = val {outer inner sublayer_size}
                }
            size
            }

        apply = inl {d with weights={weights biases dim={outer inner} streams} input} s ->
            inl span = primal input .span_outer
            assert (span = 1) "The differentiable plasticity layer supports only online learning for now."
            inl {state H} =
                match d with
                | {state} -> state
                | _ -> {
                    H=s.CudaTensor.zero_view {elem_type=float; dim=inner.weights.plastic, outer} .basic
                    state=s.CudaTensor.zero {elem_type=float; dim=span, size}
                    }

            inl apply =
                inl {alpha upper lower} = wrap_split ((), inner.biases) biases.weight
                inm data = segmented_init {dim=span,outer} {bias=const one; input=load input; state=load state}
                inl data = Struct.map' (inl data -> data.basic) data
                inm {static plastic} =
                    matmult_stream
                        {
                        static={weights with data}
                        plastic={data weight={T=H}; streams block=()}
                        }
                inl {eta out=static} = wrap_split ((), inner.weights.static) static
                inm out = map CudaAD.Activation.hebb_tanh {alpha plastic static}
                inm H = wn_hebb {H upper lower eta out input=data}
                succ {out state={state=out; H}}

            inl {out={out state} bck} = apply s
            {out state bck}
        optimize = Optimizer.kfac
        block = ()
        }
```

Yeah, this looks nice.

Let me try running it.

...6.4s. I think it was above 7 last time. Let me undo it so I can check it out.

7.6s last time.

11:35am. Yeah, it got 1.2s faster. This is roughly 15%. Well worth it."