"11:40am. It is important to have realistic expectations regarding these things. Now that I understand what Binary Pattern is about, I should not be surprised when modulation using the activities fails.
What is important is that it does not hinder the optimization process which is the case here. It WILL work on other tasks like poker.

One thing I would expect is that this single modulation RNN will behave much like the full modulation RNN which modulates all the inputs and the outputs due to the new gates. I do not think it will be explosive like the one that required -25 learning rates to be stable. That is what I was aiming for with this.

I am done with Binary Pattern for good. It is useful for telling me whether the thing works at all, but I need a real task in order to be able to tell whether plasticity is enabling any greater capabilities for the RNN compared to the vanilla.

One thing I am sure of is that these gated modulators are the decisive idea that will swing things in my favor.

12pm. Before I end this arc for good with my victory. Let me try a little thing. I will initialize the eta's bias to a positive value.

Here this will be useless, but in a realistic setting what this would do is prime the RNN to learn to learn on novel stimulus."