"9:40am. https://matrixcalc.org/en/#%7B%7B1/4,-%281/4%29,0,0%7D,%7B0,1/4,-%281/4%29,0%7D,%7B0,0,1/4,-%281/4%29%7D,%7B0,0,0,1/4%7D%7D%5E%28-1%29
```
0.25 -0.25 0     0
0    0.25  -0.25 0
0    0     0.25  -0.25
0    0     0     0.25
```

This inverts into

```
4 4 4 4
0 4 4 4
0 0 4 4
0 0 0 4
```

Amazing. So it was like this. There is a lot I do not understand about matrix inversion it seems.

Last night just like I did for the modulation activation, I strained my mind to its limit to once again figure out how to do credit assignment for the critics. I figured out one thing - I've completely overlooked the eligibility traces in Zap. I started thinking along those lines and how I could improve credit assignment and one thing is starting to become blindingly obvious that should have been obvious, but was not previously.

Algorithms like Zap are the only chance of breaking the deadlock on improving upon MC and TD. It is simply impossible to go in any other direction.

The matrix above I just inverted is on a simple MDP with 4 states (the final state is `0 0 0 0`). What `0.25 -0.25 0 0` represents is the difference of the first and the second state. This is emulating the Zap update of the steady state matrix. The same goes for the other rows, I am just going through the states forward linearly.

What the above demonstrates is the steady state matrix acts as a eligibility trace. There is no need to actually put them in explicitly it seems.

```
0.3      -0.2 0 0
0   0.3  -0.2 0
0   0    0.3  -0.2
0   0    0    0.3
```

When I invert this, so as to emulate the effect of regularization I get...

```
10/3 20/9 40/27 80/81
0    10/3 20/9  40/27
0    0    10/3  20/9
0    0    0     10/3
```

Here are eligibility traces with decay essentially. Without being put in, they arise naturally in the steady state matrix.

Ok, I am convinced without a shadow of a doubt - Zap is not a NG method. And I should be using it despite some slight weakness compared to MC. It is essentially doing model based RL, without actually having a model explicitly. That is what Zap is.

The reason why I could not draw its full power compared to MC is because the poker game I was testing it on had sequences that were too short to make any real difference.

10:15am. So now that I've concluded this, probably the way to exceed BPTT's truncation length would be to use an algorithm similar to this, though I do not know how at the moment.

10:20am. ...There is no that much leeway here though. I should be able to find it if I reason through the RNN's updates...

Alternatively I can keep using BPTT combined with Zap TD. That will give me what I want regardless.

There is one thing I need to keep in mind when using Zap. The above inversion demonstrates that the eligibility trace decays quite rapidly, so I might want to use a lower epsilon than my usual `2 ** -3`. Maybe something like `2 ** -8` at least.

10:35am. I had quite a restless night because I was obsessively thinking about this so I will take it easy today.

These insights just now make it easy for me to decide what I want to do. I will use Zap TD for the critic without back naturalization and not do MC anymore. Without regularization the above demonstrates that Zap TD would be equivalent to MC anyway.

10:45am. With TD values being feed to the actor that will minimize the variance to its limit, and the critic doing Zap TD will give me the benefit of doing model based RL. I can dispose of the concepts of eligiblity traces, their decay and discount factors as they are implicit in Zap anyway. The discount factor for example could be explained by doing EMA updates of the steady state matrix. With EMA updates the earlier updates will be decayed compared to the latest ones which is the intention behind discount factors.

10:50am. I have some more insight - if I want the benefit of Zap for feedforward linear net, I can just use the update as is and it would work.

...But if that reasoning is right, then it should also work for an RNN...should it?

Maybe for a single step update, but I am not completely sure about this.

11am. This is worth thinking about, but I'll let myself arrive at the answer gradually.

Being in this highly wound up state that I am now is not good. I am in no condition to do programming today. Usually night should be for resting, but instead I spent it all brainstorming. Yesterday, I noticed there are a lot of good references in the backpropamine paper that I'd overlooked. I want to take some time to go through that now.

11:05am. I am starting to converge to the conclusion that it would not work for a linear net internally. The reason is because it would be smearing gradients to inputs that might not result in them.

The situation is different for the linear critic - it is modeling a value function in the top layer directly. I should not let the gradients flow from them. Rather the reasoning for this should be that if the inputs to the policy layer (which the critic also gets) are poor, then the policy itself will be poor. It should be perfectly fine to infer the value of the policy at that timestep using just the inputs to it.

This was always my intention as a matter of fact.

11:20am. I cannot make money, so I'll make weapons. That is what the purpose of thinking is.

This reasoning that I am refining over time is a true weapon.

11:30am. I'll take a break soon. Right now I am thinking about Rudder. There is a definitely link between model based learning and reward redistribution. On the simple MDP made for studying the effects of variance on states that I am imagining right now, Zap's behavior should be the same as Rudder's.

11:45am. Let me finally take a break. I'll do some more matrix inversion tests after that to verify my thinking."